requirements.txt
pymodbus
minimalmodbus
python-dotenv
requests